Remove `model.strategy` module
Remove the `model.strategy` module and rewritten some functions as to
reuse some come. This was the last component of the `model` module, that
was therefore deleted too.
